Move templates to (temporary) new structure [newstructure]
Ideally each template should be under each technology dir but
the way the generic generator is written now, it assumes everything
under one directory.  For now, it can stay like this (but under
the output_filters directory).
