+++ output_filters/templates/D-Bus/Enumeration.tpl
