+++ output_filters/templates/D-Bus/Namespace.tpl
