+++ output_filters/templates/sds-bamm/sds-bamm-aspect-model.md
